fix(sk4): apply ElementKind respell + ArchSku::Sm90 workspace-wide
The sk4 pin renamed ElementKind variants and added ArchSku::Sm90; only
baracuda-cuda-emit had been migrated (an under-scoped earlier pass). Apply the
same across the remaining consumers.

- ElementKind respell (247 refs across cutlass/kernels/flashinfer): S8->I8,
  S4->I4, Bin->B1, Complex32->Complex64, Complex64->Complex128 (component-width
  -> NumPy total-width), Fp8E4M3->Fp8E4M3FN. Enum-name references only; no
  token / kernel-body / golden bytes touched (0 old names remain).
- ArchSku::Sm90 exhaustive-match coverage: nvrtc `arch_flag` gets a DISTINCT
  `Sm90 => "sm_90"` arm (portable-Hopper flag, != sm_90a); cutlass dispatch
  OR-combines `Sm90a | Sm90` (cutlass ships no kernel for either Hopper target
  -> both decline identically).

Verified: `cargo check --all-targets` (nvrtc) on baracuda-{cutlass,kernels,
flashinfer,cuda-emit} = clean; cuda-emit host goldens 288 passed;
kiss_dtype_manifest 4/4 bidirectional. NOT yet re-checked --all-targets:
baracuda-kernels-bench (depends on these; had the nvrtc Sm90 break, now fixed).

+++ crates/baracuda-kernels/src/fft/fftshift.rs
@@ -3,7 +3,7 @@
 //! cuFFT has no native shift, so these are hand-rolled kernels in
 //! `baracuda-kernels-sys`. The kernel is element-width-generic (4 / 8
 //! / 16-byte cells) so the same code covers `f32`, `f64`,
-//! [`Complex32`], and [`Complex64`] without per-type templating —
+//! [`Complex64`], and [`Complex128`] without per-type templating —
 //! shift is a pure index permutation (no arithmetic on the element
 //! values), so the element type is irrelevant beyond its byte width.
 //!
@@ -79,7 +79,7 @@ pub struct FftShiftArgs<'a, T: Element> {
 ///
 /// **Dtypes**: any [`Element`] — kernel dispatches on
 /// `size_of::<T>()` (4 / 8 / 16-byte cells), so `f32`, `f64`,
-/// `Complex32`, `Complex64` all work without per-type templating.
+/// `Complex64`, `Complex128` all work without per-type templating.
 ///
 /// **Shape**: `[batch, n]`. Out-of-place only (in-place shift would
 /// need a 2-phase swap).
@@ -111,7 +111,7 @@ impl<T: Element> FftShiftPlan<T> {
         }
         // Kernel handles cells of 4, 8, or 16 bytes — any baracuda
         // [`Element`] with one of those widths is supported. Today
-        // that's f32 / f64 / Complex32 / Complex64; rejecting the
+        // that's f32 / f64 / Complex64 / Complex128; rejecting the
         // others up front keeps the supported set narrow and obvious.
         let size = core::mem::size_of::<T>();
         if !matches!(size, 4 | 8 | 16) {
@@ -131,7 +131,7 @@ impl<T: Element> FftShiftPlan<T> {
         }
 
         let math_precision = match T::KIND {
-            ElementKind::F64 | ElementKind::Complex64 => MathPrecision::F64,
+            ElementKind::F64 | ElementKind::Complex128 => MathPrecision::F64,
             _ => MathPrecision::F32,
         };
         let precision_guarantee = PrecisionGuarantee {
